Increase timeouts and parallel connections to resource-server
We had some issues with too much parallel request to the auth-server. It
turned out that requests to the resource-server, which are necessary to
validate the user's credentials, queue up in the HTTP connection pool
queue, which leads to timeouts of the auth-server client's access token,
making the request fail with "Your token is invalid" messages.